Modifiy Seeded Data
Modifiy Seeded Data

+++ HealthBanc/Controllers/BackendAdmin/BackendAdminAuthController.cs
@@ -126,7 +126,7 @@ namespace HealthBanc.Controllers
         [ProducesResponseType(401, Type = typeof(ResponseMessage))]
         [ProducesResponseType(400, Type = typeof(ResponseMessage))]
         [ProducesResponseType(404, Type = typeof(ResponseMessage))]
-        [Authorize(Roles = "Super-Administrator")]
+        [Authorize(Roles = "Super-Administrator,Administrator,Technical-Support,Analyst")]
         [HttpPost("[action]")]
         public async Task<IActionResult> CreateBackendAdmin(CreateAdminViewModel createAdminViewModel)
         {
@@ -137,55 +137,64 @@ namespace HealthBanc.Controllers
                 string email = User.FindFirst(ClaimTypes.Email)?.Value;
                 int Id = int.Parse(userId);
 
-                // get logged in user
-                var loggedinuser = await _userRepository.GetByEmailAsync(email);
-                //get logged in admin mail
-                string loggedInAdminMail = email;
-                // remove the .admin from the userMail
-                var newLoggedInAdminMail = loggedInAdminMail.Remove(loggedInAdminMail.Length - 6);
-
-                if (!createAdminViewModel.Email.EndsWith("@sterling.ng"))
+                var checkRole = User.IsInRole("Super-Administrator");
+                if (checkRole)
                 {
-                    return BadRequest(new ResponseMessage { Message = "Email is not a valid sterling email" });
-                }
+                    // get logged in user
+                    var loggedinuser = await _userRepository.GetByEmailAsync(email);
+                    //get logged in admin mail
+                    string loggedInAdminMail = email;
+                    // remove the .admin from the userMail
+                    var newLoggedInAdminMail = loggedInAdminMail.Remove(loggedInAdminMail.Length - 6);
 
-                var checkEmail = await _userManager.FindByEmailAsync($"{createAdminViewModel.Email}.admin");
-                if (checkEmail != null) return BadRequest(new ResponseMessage{ Message = "Email Already Exist" });
-                var checkIfUserExist = await _userRepository.FindByUniqueUsername(createAdminViewModel.UserName);
-                if (checkIfUserExist != null) return BadRequest(new ResponseMessage { Message = "Username Already Exist" });
-
-                var backedAdmin = await _adminRepository.GetAdminByEmail(newLoggedInAdminMail);
-
-                var admin = new ApplicationUser()
-                {
-                    FirstName = createAdminViewModel.FirstName,
-                    LastName = createAdminViewModel.LastName,
-                    Email = createAdminViewModel.Email+ ".admin",
-                    UserName = createAdminViewModel.Email+ ".admin",
-                    UniqueUsername = createAdminViewModel.UserName,
-                    EmailConfirmed = true
-                };
-                var result = _userManager.CreateAsync(admin).Result;
-                if (result.Succeeded)
-                {
-                    var role = await _roleRepository.GetRole(createAdminViewModel.RoleId);
-                    await _userManager.AddToRoleAsync(admin, role.Name);
-                    BackendAdminUser adminUser = new BackendAdminUser()
+                    if (!createAdminViewModel.Email.EndsWith("@sterling.ng"))
                     {
-                        Email = createAdminViewModel.Email,
+                        return BadRequest(new ResponseMessage { Message = "Email is not a valid sterling email" });
+                    }
+
+                    var checkEmail = await _userManager.FindByEmailAsync($"{createAdminViewModel.Email}.admin");
+                    if (checkEmail != null) return BadRequest(new ResponseMessage { Message = "Email Already Exist" });
+                    var checkIfUserExist = await _userRepository.FindByUniqueUsername(createAdminViewModel.UserName);
+                    if (checkIfUserExist != null) return BadRequest(new ResponseMessage { Message = "Username Already Exist" });
+
+                    var backedAdmin = await _adminRepository.GetAdminByEmail(newLoggedInAdminMail);
+
+                    var admin = new ApplicationUser()
+                    {
                         FirstName = createAdminViewModel.FirstName,
                         LastName = createAdminViewModel.LastName,
-                        ClassOrRoleId = createAdminViewModel.RoleId
+                        Email = createAdminViewModel.Email + ".admin",
+                        UserName = createAdminViewModel.Email + ".admin",
+                        UniqueUsername = createAdminViewModel.UserName,
+                        EmailConfirmed = true
                     };
-                    _adminRepository.Create(adminUser);
-                    await _adminRepository.Save();
+                    var result = _userManager.CreateAsync(admin).Result;
+                    if (result.Succeeded)
+                    {
+                        var role = await _roleRepository.GetRole(createAdminViewModel.RoleId);
+                        await _userManager.AddToRoleAsync(admin, role.Name);
+                        BackendAdminUser adminUser = new BackendAdminUser()
+                        {
+                            Email = createAdminViewModel.Email,
+                            FirstName = createAdminViewModel.FirstName,
+                            LastName = createAdminViewModel.LastName,
+                            ClassOrRoleId = createAdminViewModel.RoleId
+                        };
+                        _adminRepository.Create(adminUser);
+                        await _adminRepository.Save();
 
-                    var auditViewModel = new AdminAuditLogViewModel(Id,backedAdmin.Id,$"{loggedinuser.UniqueUsername} added {adminUser.Email}",ServiceNames.HealthBanc.ToString());
-                    BackgroundJob.Enqueue(() => _auditLogServices.AdminCreateAuditLog(auditViewModel));
+                        var auditViewModel = new AdminAuditLogViewModel(Id, backedAdmin.Id, $"{loggedinuser.UniqueUsername} added {adminUser.Email}", ServiceNames.HealthBanc.ToString());
+                        BackgroundJob.Enqueue(() => _auditLogServices.AdminCreateAuditLog(auditViewModel));
 
-                    return Ok(new ResponseMessage{ Message = "Admin has been created successfully", Status = true });
+                        return Ok(new ResponseMessage { Message = "Admin has been created successfully", Status = true });
+                    }
+                    return BadRequest(new ResponseMessage { Message = result.Errors.FirstOrDefault().Description.ToString() });
                 }
-                return BadRequest(new ResponseMessage { Message = result.Errors.FirstOrDefault().Description.ToString() });
+                else
+                {
+                    return BadRequest(new ResponseMessage { Message = "You do not have the authority to add a new admin,contact the Super Admin" });
+                }
+                
             }
             //return validation errors
             var errors = new List<string>();
@@ -229,34 +238,39 @@ namespace HealthBanc.Controllers
             int Id = int.Parse(userId);
             var loggedInUser = await _userRepository.FindByIdAsync(Id);
 
-            //get logged in admin mail
-            string loggedInAdminMail = User.FindFirst(ClaimTypes.Email)?.Value;
-            // remove the .admin from the userMail
-            var newLoggedInAdminMail = loggedInAdminMail.Remove(loggedInAdminMail.Length - 6);
-            var backedAdmin = await _adminRepository.GetAdminByEmail(newLoggedInAdminMail);
-
-            var adminUserMail = email + ".admin";
-            var user = await _userManager.FindByEmailAsync(adminUserMail);
-            var admin = await _adminRepository.GetAdminByEmail(email);
-            if(user != null)
+            var checkRole = User.IsInRole("Super-Administrator");
+            if (checkRole)
             {
-                var userRole = await _userManager.GetRolesAsync(user);
-                var removeRoleResult = _userManager.RemoveFromRoleAsync(user, userRole.FirstOrDefault()).Result;
-                var role = await _roleRepository.GetRole(roleId);
-                var result = _userManager.AddToRoleAsync(user, role.Name).Result;
-                if (result.Succeeded)
+                //get logged in admin mail
+                string loggedInAdminMail = User.FindFirst(ClaimTypes.Email)?.Value;
+                // remove the .admin from the userMail
+                var newLoggedInAdminMail = loggedInAdminMail.Remove(loggedInAdminMail.Length - 6);
+                var backedAdmin = await _adminRepository.GetAdminByEmail(newLoggedInAdminMail);
+
+                var adminUserMail = email + ".admin";
+                var user = await _userManager.FindByEmailAsync(adminUserMail);
+                var admin = await _adminRepository.GetAdminByEmail(email);
+                if (user != null)
                 {
-                    admin.ClassOrRoleId = roleId;
-                    _adminRepository.Update(admin);
-                    await _adminRepository.Save();
+                    var userRole = await _userManager.GetRolesAsync(user);
+                    var removeRoleResult = _userManager.RemoveFromRoleAsync(user, userRole.FirstOrDefault()).Result;
+                    var role = await _roleRepository.GetRole(roleId);
+                    var result = _userManager.AddToRoleAsync(user, role.Name).Result;
+                    if (result.Succeeded)
+                    {
+                        admin.ClassOrRoleId = roleId;
+                        _adminRepository.Update(admin);
+                        await _adminRepository.Save();
 
-                    var auditViewModel = new AdminAuditLogViewModel(Id, backedAdmin.Id, $"{loggedInUser.UniqueUsername} changed {email} role", ServiceNames.HealthBanc.ToString());
-                    BackgroundJob.Enqueue(() => _auditLogServices.AdminCreateAuditLog(auditViewModel));
+                        var auditViewModel = new AdminAuditLogViewModel(Id, backedAdmin.Id, $"{loggedInUser.UniqueUsername} changed {email} role", ServiceNames.HealthBanc.ToString());
+                        BackgroundJob.Enqueue(() => _auditLogServices.AdminCreateAuditLog(auditViewModel));
 
-                    return Ok(new ResponseMessage {Message="Role was changed successfully", Status=true });
+                        return Ok(new ResponseMessage { Message = "Role was changed successfully", Status = true });
+                    }
                 }
+                return NotFound(new ResponseMessage { Message = "User does not exist" });
             }
-            return NotFound(new ResponseMessage { Message="User does not exist" });
+            return BadRequest(new ResponseMessage { Message = "You do not have the authority to change admin role,contact the Super Admin" });
         }
 
         //WORKING1
@@ -288,32 +302,36 @@ namespace HealthBanc.Controllers
             int Id = int.Parse(userId);
             var loggedInUser = await _userRepository.FindByIdAsync(Id);
 
-
-            //get logged in admin mail
-            string loggedInAdminMail = User.FindFirst(ClaimTypes.Email)?.Value;
-            // remove the .admin from the userMail
-            var newLoggedInAdminMail = loggedInAdminMail.Remove(loggedInAdminMail.Length - 6);
-            var backedAdmin = await _adminRepository.GetAdminByEmail(newLoggedInAdminMail);
-
-            var adminUserMail = email + ".admin";
-            var user = await _userManager.FindByEmailAsync(adminUserMail);
-            if (user != null)
+            var checkRole = User.IsInRole("Super-Administrator");
+            if (checkRole)
             {
-                var result = _userManager.DeleteAsync(user).Result;
-                if (result.Succeeded)
+                //get logged in admin mail
+                string loggedInAdminMail = User.FindFirst(ClaimTypes.Email)?.Value;
+                // remove the .admin from the userMail
+                var newLoggedInAdminMail = loggedInAdminMail.Remove(loggedInAdminMail.Length - 6);
+                var backedAdmin = await _adminRepository.GetAdminByEmail(newLoggedInAdminMail);
+
+                var adminUserMail = email + ".admin";
+                var user = await _userManager.FindByEmailAsync(adminUserMail);
+                if (user != null)
                 {
-                    var admin = await _adminRepository.GetAdminByEmail(email);
-                    _adminRepository.Delete(admin);
-                    await _adminRepository.Save();
+                    var result = _userManager.DeleteAsync(user).Result;
+                    if (result.Succeeded)
+                    {
+                        var admin = await _adminRepository.GetAdminByEmail(email);
+                        _adminRepository.Delete(admin);
+                        await _adminRepository.Save();
 
-                    var auditViewModel = new AdminAuditLogViewModel(Id, backedAdmin.Id, $"{loggedInUser.UniqueUsername} removed {email}", ServiceNames.HealthBanc.ToString());
-                    BackgroundJob.Enqueue(() => _auditLogServices.AdminCreateAuditLog(auditViewModel));
+                        var auditViewModel = new AdminAuditLogViewModel(Id, backedAdmin.Id, $"{loggedInUser.UniqueUsername} removed {email}", ServiceNames.HealthBanc.ToString());
+                        BackgroundJob.Enqueue(() => _auditLogServices.AdminCreateAuditLog(auditViewModel));
 
-                    return Ok(new ResponseMessage { Message = "Admin was deleted successfully", Status = true });
+                        return Ok(new ResponseMessage { Message = "Admin was deleted successfully", Status = true });
+                    }
+                    return BadRequest("An error occurred while trying to change to delete admin");
                 }
-                return BadRequest("An error occurred while trying to change to delete admin");
+                return NotFound(new ResponseMessage { Message = "User does not exist" });
             }
-            return NotFound(new ResponseMessage { Message = "User does not exist" });
+            return BadRequest(new ResponseMessage { Message = "You do not have the authority to remove an admin,contact the Super Admin" });
         }        
     }
 }
